task(SDK-5584) - Removes InAppMediaDelegate.kt
