+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/media/InAppMediaConfig.kt
@@ -1,0 +1,8 @@
+package com.clevertap.android.sdk.inapp.media
+
+internal data class InAppMediaConfig(
+    val imageViewId: Int,
+    val clickableMedia: Boolean,
+    val useOrientationForImage: Boolean = true,
+    val videoFrameId: Int = 0,
+)
